EmojiBot-py/requirements.txt
discord.py
webcolors
asyncio